Add defensive guards against sub-equal strings
ART trees can't store strings where one is a prefix of another (but
isn't that the point of a radix tree?). The suggested fix is to
terminiate all input strings with a null byte so "santa" doesn't become
a prefix match of "santaana", so you'd store "santa\0" "santaana\0"
which doesn't bust the algorithm.

This commit makes fake null terminators at the end of strings when
applicable. ymmv.