Cherry pick "fix a bug in range scan" from origin
This is just
https://github.com/armon/libart/commit/b443d325461b2eed167adf77ca9660adf5688611

Not entirely sure why it is necessary, but since it got included
upstream we're including it here (but then they reverted it).

All tests passed before this patch and after this patch, but it would be
nice to know what was broken before the patch I guess.

Looking further this is from issue here where they declared the fix
incorrect and reverted it anyway:
https://github.com/armon/libart/pull/37